Added serial port factory
